Add tasks to update forecasts and send mail
Forecasts should be updated daily and then email messages should be sent to each
registered user with the current forecast. These tasks can be scheduled with
cron or another scheduling service.

Add a task to update forecasts for each ZIP code belonging to a registered
users.

Add a task to send the current forecast via email to each registered user.