There is a file (dino_class) that defines the dino objects.  The dino objects hold the attributes of the dinos.  To initialize a dino object, pass in a dictionary that contains the data as values associated with keys that match the headers in the csv file.
There is also a dino_filters class.  This class allows us to create
an object that is a list of dino objects.  This class also contains
all of the filter methods that are used to select the dinos that
fall into a certain category.  The filter functions in the class
return the filtered list as a dino_filters object.  I did this
so that I can chain methods together and then print the final
result (also through method chaining).

The main function reads the csv files and creates the list of
dino objects and then the dino_filter object.  Then I run
a bunch of tests and returns the dino_filter object to the
user so that they can also do more queries.