Reorganized overlord to redirect to index page after each request and use flash to decrease number of views.
Overlord now has an index function that determines which page should
be served to the user based on the status of the bomb.
If there was an error, the get/post method will put the error
message into the flash dictionary and will be displayed after
fetching the new page that should be displayed.

I also added a /newbomb get method so that I could create a
new bomb without restarting the web server.  Restarting the bomb
used to take place by going back to the index route, but because
that route is visited after all routes, I could not leave the
bomb reboot in that method.

I also added a 'before do' and 'after do' to get the current
bomb object and update the current bomb object in the session
variable.