Third Stab at Cucumber Rules
It would be great if someone could sit with me for 30 minutes or so to
help me understand what level of detail to include in these specs. In
some cases it seems like I’m too detailed and in other cases I’m not
specific enough. Does Enova have any published guidelines on this?