Refactored API library call.  Simplified the parse_response function so that if an error was not generated, an Array is always returned (rather than and Array if there are multiple books and a dictionary if there is only one list).
Added a code block to resource :book in route file to make
it possible to access urls as /book/search  and /book/results
rather than having to use '/book/new/search'.  I was doing this
because of the route for the #show function included in resource.

I also changed the names of two of the controller functions
and their matching views
get_user_query --> search
perform_query --> results