Started working on fixing hound issues
I started to work on fixing the hound issues (for paths.rb at least) related to cucumber-rails. Came to the conclusion that this might not be required, so I'm going to wait to refactor the cucumber-rails stuff. I also cleaned up overlord.rb quite a bit as well as added an executeable run_all_tests.sh which runs the rspec tests as well as the cucumber tests.