Major Restructuring
I majorly restructured the app. I split all graph generation out of the artist model. This was massively illogical to do in the first place. (Artists do not need to know anything about graphs!).
Now graph generation is handled by it's own model (which also caches each graph based on the depth).

I also did some minor refactoring throughout the project in the hopes that the code would be more concise and understandable.